rename higlights screen file and folder
